+++ setup.cfg
@@ -2,13 +2,14 @@
 name = UzbekStemmer
 version = 0.2.0
 author = Maksud Sharipov, Ulugbek Salaev, Allabergan Yuldashev, Jasur Sobirov
+author_email = maqsbek72@gmail.com
 author_email = ulugbek0302@gmail.com
 description = Uzbek Stemmer for Python
 long_description = file: README.md
 long_description_content_type = The Uzbek stemming algorithm was created by [Maksud Sharipov, Ulugbek Salaev, Allabergan Yuldashev, Jasur Sobirov]. It stems all Uzbek words and it is not included any lexicon.
 url = https://github.com/UlugbekSalaev/UzbekStemmer
 project_urls =
-    Bug Tracker = https://github.com/UlugbekSalaev/UzbekStemmer
+    Bug Tracker = https://github.com/MaksudSharipov/UzbekStemmer
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
